feat: add suvx support
this commit adds a new binary 'suvx' which is a drop-in replacement for 'uvx', and
it changes the structure of this project and renames the project to 'semcp' stands for
'secure execution of MCP'. now the repo is repurposed to provide a tool kit for running
MCP servers locally in docker containers.

the common code is moved to the 'common' crate and the 'snpx' and 'suvx' are the two main
executables.

Signed-off-by: Jiaxiao Zhou <duibao55328@gmail.com>